Test factory with unit testing
